Remove hard restriction on AbstractReference inheritance
Due to buggy nature of how disalow subclassing works, it's better to
remove it rather then keep it and factor complexity and ambiguite in.

From now while AbstractReference may not be subclassed directrly, it
technically can, but it will have no effect and it will not handle any
references at all and will not track object destroying.